refactor(desktop): extract browser HTTP service with all 16 routes
Move the localhost HTTP server's route handler (handleDesktopBrowserServiceRequest
+ 7 helpers) into browser-pane/http-service.ts. The module owns:

- request header parsers (token / workspace / session / space)
- JSON read/write helpers
- serializeEvalResult
- handleRequest with all 16 routes:
  * GET /health, /tabs, /downloads, /console, /errors, /requests,
    /requests/<id>, /cookies, /page, /operator-surface-context
  * POST /navigate, /tabs, /tabs/select, /tabs/close, /cookies,
    /evaluate, /context-click, /mouse, /keyboard, /screenshot

Server lifecycle (start/stop, auth-token rotation, capability config
sync) stays in main.ts because it's tied to runtimeStatus and
emitRuntimeState. main.ts now wires a BrowserHttpService instance and
the createServer callback delegates to it.

main.ts -976. Typecheck clean.
