+++ desktop/electron/main.ts
@@ -160,6 +160,10 @@ import {
   createBrowserPaneTabState,
   type BrowserPaneTabState,
 } from "./browser-pane/tab-state.js";
+import {
+  createBrowserHttpService,
+  type BrowserHttpService,
+} from "./browser-pane/http-service.js";
 import type {
   BrowserCopyWorkspaceProfilePayload,
   BrowserImportProfilePayload,
@@ -5242,76 +5246,17 @@ async function syncDesktopBrowserCapabilityConfig(): Promise<void> {
   );
 }
 
-function desktopBrowserServiceTokenFromRequest(
-  request: IncomingMessage,
-): string {
-  const raw = request.headers["x-holaboss-desktop-token"];
-  if (Array.isArray(raw)) {
-    return (raw[0] || "").trim();
-  }
-  return typeof raw === "string" ? raw.trim() : "";
-}
+// desktopBrowserServiceTokenFromRequest moved to browser-pane/tab-state.ts.
 
-function desktopBrowserWorkspaceIdFromRequest(
-  request: IncomingMessage,
-): string {
-  const raw = request.headers["x-holaboss-workspace-id"];
-  if (Array.isArray(raw)) {
-    return (raw[0] || "").trim();
-  }
-  return typeof raw === "string" ? raw.trim() : "";
-}
+// desktopBrowserWorkspaceIdFromRequest moved to browser-pane/tab-state.ts.
 
-function desktopBrowserSessionIdFromRequest(
-  request: IncomingMessage,
-): string {
-  const raw = request.headers["x-holaboss-session-id"];
-  if (Array.isArray(raw)) {
-    return (raw[0] || "").trim();
-  }
-  return typeof raw === "string" ? raw.trim() : "";
-}
+// desktopBrowserSessionIdFromRequest moved to browser-pane/tab-state.ts.
 
-function desktopBrowserSpaceFromRequest(
-  request: IncomingMessage,
-): BrowserSpaceId {
-  const raw = request.headers["x-holaboss-browser-space"];
-  if (Array.isArray(raw)) {
-    return browserSpaceId(raw[0] || "", "agent");
-  }
-  return browserSpaceId(typeof raw === "string" ? raw.trim() : "", "agent");
-}
+// desktopBrowserSpaceFromRequest moved to browser-pane/tab-state.ts.
 
-function writeBrowserServiceJson(
-  response: ServerResponse<IncomingMessage>,
-  statusCode: number,
-  payload: unknown,
-): void {
-  response.statusCode = statusCode;
-  response.setHeader("content-type", "application/json; charset=utf-8");
-  response.end(`${JSON.stringify(payload)}\n`);
-}
+// writeBrowserServiceJson moved to browser-pane/tab-state.ts.
 
-async function readBrowserServiceJsonBody(
-  request: IncomingMessage,
-): Promise<Record<string, unknown>> {
-  const chunks: Buffer[] = [];
-  for await (const chunk of request) {
-    chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
-  }
-  if (chunks.length === 0) {
-    return {};
-  }
-  const raw = Buffer.concat(chunks).toString("utf-8").trim();
-  if (!raw) {
-    return {};
-  }
-  const parsed = JSON.parse(raw) as unknown;
-  if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
-    throw new Error("Request body must be a JSON object.");
-  }
-  return parsed as Record<string, unknown>;
-}
+// readBrowserServiceJsonBody moved to browser-pane/tab-state.ts.
 
 // browserPagePayload moved to browser-pane/tab-state.ts.
 
@@ -5491,27 +5436,7 @@ function operatorSurfaceContextPayload(workspaceId: string): OperatorSurfaceCont
   };
 }
 
-function serializeBrowserEvalResult(value: unknown): unknown {
-  if (value === undefined) {
-    return null;
-  }
-  if (
-    value === null ||
-    typeof value === "string" ||
-    typeof value === "number" ||
-    typeof value === "boolean"
-  ) {
-    return value;
-  }
-  if (typeof value === "bigint") {
-    return value.toString();
-  }
-  try {
-    return JSON.parse(JSON.stringify(value)) as unknown;
-  } catch {
-    return String(value);
-  }
-}
+// serializeBrowserEvalResult moved to browser-pane/tab-state.ts.
 
 function isAbortedBrowserLoadError(error: unknown): boolean {
   return isAbortedBrowserLoadErrorUtil(error);
@@ -5526,906 +5451,7 @@ function isAbortedBrowserLoadFailure(
 
 // navigateActiveBrowserTab moved to browser-pane/tab-state.ts.
 
-async function handleDesktopBrowserServiceRequest(
-  request: IncomingMessage,
-  response: ServerResponse<IncomingMessage>,
-): Promise<void> {
-  try {
-    const requestUrl = new URL(request.url || "/", "http://127.0.0.1");
-    const pathname = requestUrl.pathname;
-    const method = (request.method || "GET").toUpperCase();
-    const targetSpace = desktopBrowserSpaceFromRequest(request);
-    const requestedWorkspaceId = desktopBrowserWorkspaceIdFromRequest(request);
-    const requestedSessionId = desktopBrowserSessionIdFromRequest(request);
-    const targetWorkspaceId = requestedWorkspaceId || activeBrowserWorkspaceId;
-
-    if (
-      !desktopBrowserServiceAuthToken ||
-      desktopBrowserServiceTokenFromRequest(request) !==
-        desktopBrowserServiceAuthToken
-    ) {
-      writeBrowserServiceJson(response, 401, { error: "Unauthorized." });
-      return;
-    }
-
-    if (method === "GET" && pathname === "/api/v1/browser/health") {
-      writeBrowserServiceJson(response, 200, { ok: true });
-      return;
-    }
-
-    if (!targetWorkspaceId) {
-      writeBrowserServiceJson(response, 409, {
-        error: "No active browser workspace is available.",
-      });
-      return;
-    }
-
-    const ensuredSessionId = targetSpace === "agent" ? requestedSessionId : null;
-    const ensureTargetBrowserSpace = async (
-      reason: string,
-    ): Promise<BrowserWorkspaceState | null> => {
-      const workspace = await ensureBrowserWorkspace(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-      );
-      if (!workspace) {
-        return null;
-      }
-      if (targetSpace === "user") {
-        if (!requestedSessionId) {
-          writeBrowserServiceJson(response, 400, {
-            error:
-              "Header 'x-holaboss-session-id' is required when targeting the user browser.",
-          });
-          return null;
-        }
-        const lockResult = ensureUserBrowserLock(
-          targetWorkspaceId,
-          requestedSessionId,
-          reason,
-        );
-        if (!lockResult.ok) {
-          writeBrowserServiceJson(response, 409, {
-            error: "User browser is locked by another agent session.",
-            code: "user_browser_locked",
-            lock_holder_session_id: lockResult.lockHolderSessionId,
-          });
-          return null;
-        }
-      } else if (requestedSessionId) {
-        touchAgentSessionBrowserSpace(targetWorkspaceId, requestedSessionId);
-      }
-      return workspace;
-    };
-
-    if (method === "GET" && pathname === "/api/v1/browser/tabs") {
-      const workspace = await ensureTargetBrowserSpace("tabs");
-      if (!workspace) {
-        return;
-      }
-      writeBrowserServiceJson(
-        response,
-        200,
-        browserWorkspaceSnapshot(targetWorkspaceId, targetSpace, ensuredSessionId, {
-          useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId,
-        }),
-      );
-      return;
-    }
-
-    if (method === "GET" && pathname === "/api/v1/browser/downloads") {
-      const workspace = await ensureTargetBrowserSpace("downloads");
-      if (!workspace) {
-        return;
-      }
-      writeBrowserServiceJson(response, 200, {
-        downloads: workspace.downloads,
-      });
-      return;
-    }
-
-    if (method === "GET" && pathname === "/api/v1/browser/console") {
-      if (!(await ensureTargetBrowserSpace("console"))) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-      const requestedLevel = requestUrl.searchParams.get("level")?.trim() || "";
-      const minimumLevel =
-        requestedLevel === "debug" ||
-        requestedLevel === "info" ||
-        requestedLevel === "warning" ||
-        requestedLevel === "error"
-          ? requestedLevel
-          : null;
-      const filtered = [...activeTab.consoleEntries].filter((entry) =>
-        minimumLevel
-          ? browserConsoleLevelRank(entry.level) >=
-            browserConsoleLevelRank(minimumLevel)
-          : true,
-      );
-      const limit = browserObservabilityLimit(
-        requestUrl.searchParams.get("limit"),
-      );
-      const entries = filtered.slice(-limit).reverse();
-      writeBrowserServiceJson(response, 200, {
-        entries,
-        total: filtered.length,
-        truncated: filtered.length > entries.length,
-      });
-      return;
-    }
-
-    if (method === "GET" && pathname === "/api/v1/browser/errors") {
-      if (!(await ensureTargetBrowserSpace("errors"))) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-      const requestedSource = browserObservedErrorSource(
-        requestUrl.searchParams.get("source")?.trim() || null,
-      );
-      const filtered = [...activeTab.errorEntries].filter((entry) =>
-        requestedSource ? entry.source === requestedSource : true,
-      );
-      const limit = browserObservabilityLimit(
-        requestUrl.searchParams.get("limit"),
-      );
-      const errors = filtered.slice(-limit).reverse();
-      writeBrowserServiceJson(response, 200, {
-        errors,
-        total: filtered.length,
-        truncated: filtered.length > errors.length,
-      });
-      return;
-    }
-
-    if (method === "GET" && pathname === "/api/v1/browser/requests") {
-      if (!(await ensureTargetBrowserSpace("requests"))) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-      const requestedResourceType =
-        requestUrl.searchParams.get("resource_type")?.trim().toLowerCase() || "";
-      const failuresOnly = requestUrl.searchParams.get("failures_only") === "true";
-      const filtered = [...activeTab.requestOrder]
-        .map((requestId) => activeTab.requests.get(requestId) ?? null)
-        .filter((request): request is BrowserRequestRecord => Boolean(request))
-        .filter((request) =>
-          requestedResourceType
-            ? request.resourceType.toLowerCase() === requestedResourceType
-            : true,
-        )
-        .filter((request) => (failuresOnly ? browserRequestFailure(request) : true));
-      const limit = browserObservabilityLimit(
-        requestUrl.searchParams.get("limit"),
-      );
-      const requests = filtered.slice(-limit).reverse().map((request) =>
-        browserRequestSummary(request),
-      );
-      writeBrowserServiceJson(response, 200, {
-        requests,
-        total: filtered.length,
-        truncated: filtered.length > requests.length,
-      });
-      return;
-    }
-
-    if (
-      method === "GET" &&
-      pathname.startsWith("/api/v1/browser/requests/")
-    ) {
-      if (!(await ensureTargetBrowserSpace("request"))) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-      const requestId = decodeURIComponent(
-        pathname.slice("/api/v1/browser/requests/".length),
-      ).trim();
-      if (!requestId) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Request id is required.",
-        });
-        return;
-      }
-      const requestRecord = activeTab.requests.get(requestId) ?? null;
-      if (!requestRecord) {
-        writeBrowserServiceJson(response, 404, {
-          error: "Browser request not found for the active tab.",
-        });
-        return;
-      }
-      writeBrowserServiceJson(response, 200, {
-        request: requestRecord,
-      });
-      return;
-    }
-
-    if (method === "GET" && pathname === "/api/v1/browser/cookies") {
-      const workspace = await ensureTargetBrowserSpace("cookies:get");
-      if (!workspace) {
-        return;
-      }
-      const useVisibleAgentSession = targetSpace === "agent" && !requestedSessionId;
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession },
-      );
-      const requestedUrl = requestUrl.searchParams.get("url")?.trim() || "";
-      const requestedName = requestUrl.searchParams.get("name")?.trim() || "";
-      const requestedDomain =
-        requestUrl.searchParams.get("domain")?.trim() || "";
-      try {
-        const cookies = await workspace.session.cookies.get({
-          ...(requestedUrl || activeTab?.view.webContents.getURL()
-            ? { url: requestedUrl || activeTab?.view.webContents.getURL() || "" }
-            : {}),
-          ...(requestedName ? { name: requestedName } : {}),
-          ...(requestedDomain ? { domain: requestedDomain } : {}),
-        });
-        writeBrowserServiceJson(response, 200, {
-          cookies: cookies.map((cookie) => ({
-            name: cookie.name || "",
-            value: cookie.value || "",
-            domain: cookie.domain || "",
-            path: cookie.path || "/",
-            secure: Boolean(cookie.secure),
-            httpOnly: Boolean(cookie.httpOnly),
-            session: Boolean(cookie.session),
-            sameSite: cookie.sameSite || "unspecified",
-            expirationDate:
-              typeof cookie.expirationDate === "number" &&
-              Number.isFinite(cookie.expirationDate)
-                ? cookie.expirationDate
-                : null,
-          })),
-        });
-      } catch (error) {
-        writeBrowserServiceJson(response, 400, {
-          error:
-            error instanceof Error
-              ? error.message
-              : "Failed to read browser cookies.",
-        });
-      }
-      return;
-    }
-
-    if (method === "GET" && pathname === "/api/v1/browser/page") {
-      const workspace = await ensureTargetBrowserSpace("page");
-      if (!workspace) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-      syncBrowserState(targetWorkspaceId, activeTab.state.id, targetSpace, ensuredSessionId);
-      writeBrowserServiceJson(response, 200, browserPagePayload(activeTab));
-      return;
-    }
-
-    if (method === "GET" && pathname === "/api/v1/browser/operator-surface-context") {
-      await ensureBrowserWorkspace(targetWorkspaceId, "user");
-      await ensureBrowserWorkspace(targetWorkspaceId, "agent");
-      writeBrowserServiceJson(response, 200, operatorSurfaceContextPayload(targetWorkspaceId));
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/navigate") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const targetUrl =
-        typeof payload.url === "string" ? payload.url.trim() : "";
-      if (!targetUrl) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Field 'url' is required.",
-        });
-        return;
-      }
-      const workspace = await ensureTargetBrowserSpace("navigate");
-      if (!workspace) {
-        return;
-      }
-      if (targetWorkspaceId && targetWorkspaceId === activeBrowserWorkspaceId) {
-        emitWorkbenchOpenBrowser({
-          workspaceId: targetWorkspaceId,
-          url: targetUrl,
-          space: targetSpace,
-          sessionId: requestedSessionId || null,
-        });
-      }
-      const snapshot = await navigateActiveBrowserTab(
-        targetWorkspaceId,
-        targetUrl,
-        targetSpace,
-        ensuredSessionId,
-      );
-      writeBrowserServiceJson(response, 200, snapshot);
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/tabs/select") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const tabId =
-        typeof payload.tab_id === "string" && payload.tab_id.trim()
-          ? payload.tab_id.trim()
-          : typeof payload.tabId === "string" && payload.tabId.trim()
-            ? payload.tabId.trim()
-            : "";
-      if (!tabId) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Field 'tab_id' is required.",
-        });
-        return;
-      }
-      const workspace = await ensureTargetBrowserSpace("tabs:select");
-      if (!workspace) {
-        return;
-      }
-      const snapshot = await setActiveBrowserTab(tabId, {
-        workspaceId: workspace.workspaceId,
-        space: targetSpace,
-        sessionId: ensuredSessionId,
-        useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId,
-      });
-      writeBrowserServiceJson(response, 200, snapshot);
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/tabs/close") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const tabId =
-        typeof payload.tab_id === "string" && payload.tab_id.trim()
-          ? payload.tab_id.trim()
-          : typeof payload.tabId === "string" && payload.tabId.trim()
-            ? payload.tabId.trim()
-            : "";
-      if (!tabId) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Field 'tab_id' is required.",
-        });
-        return;
-      }
-      const workspace = await ensureTargetBrowserSpace("tabs:close");
-      if (!workspace) {
-        return;
-      }
-      const snapshot = await closeBrowserTab(tabId, {
-        workspaceId: workspace.workspaceId,
-        space: targetSpace,
-        sessionId: ensuredSessionId,
-        useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId,
-      });
-      writeBrowserServiceJson(response, 200, snapshot);
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/cookies") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const workspace = await ensureTargetBrowserSpace("cookies:set");
-      if (!workspace) {
-        return;
-      }
-      const useVisibleAgentSession = targetSpace === "agent" && !requestedSessionId;
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession },
-      );
-      const activeTabUrl = activeTab?.view.webContents.getURL()?.trim() || "";
-      const targetUrl =
-        (typeof payload.url === "string" && payload.url.trim()
-          ? payload.url.trim()
-          : activeTabUrl) || "";
-      const cookieName =
-        typeof payload.name === "string" && payload.name.trim()
-          ? payload.name.trim()
-          : "";
-      if (!targetUrl) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Field 'url' is required when there is no active browser page.",
-        });
-        return;
-      }
-      if (!cookieName) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Field 'name' is required.",
-        });
-        return;
-      }
-      try {
-        const expirationDate =
-          typeof payload.expiration_date === "number" &&
-          Number.isFinite(payload.expiration_date)
-            ? payload.expiration_date
-            : typeof payload.expirationDate === "number" &&
-                Number.isFinite(payload.expirationDate)
-              ? payload.expirationDate
-              : undefined;
-        await workspace.session.cookies.set({
-          url: targetUrl,
-          name: cookieName,
-          value: typeof payload.value === "string" ? payload.value : "",
-          ...(typeof payload.domain === "string" && payload.domain.trim()
-            ? { domain: payload.domain.trim() }
-            : {}),
-          ...(typeof payload.path === "string" && payload.path.trim()
-            ? { path: payload.path.trim() }
-            : {}),
-          ...(typeof payload.secure === "boolean"
-            ? { secure: payload.secure }
-            : {}),
-          ...(typeof payload.http_only === "boolean"
-            ? { httpOnly: payload.http_only }
-            : typeof payload.httpOnly === "boolean"
-              ? { httpOnly: payload.httpOnly }
-              : {}),
-          ...(payload.same_site === "unspecified" ||
-          payload.same_site === "no_restriction" ||
-          payload.same_site === "lax" ||
-          payload.same_site === "strict"
-            ? { sameSite: payload.same_site }
-            : payload.sameSite === "unspecified" ||
-                payload.sameSite === "no_restriction" ||
-                payload.sameSite === "lax" ||
-                payload.sameSite === "strict"
-              ? { sameSite: payload.sameSite }
-              : {}),
-          ...(typeof expirationDate === "number" ? { expirationDate } : {}),
-        });
-        await workspace.session.cookies.flushStore();
-        const cookies = await workspace.session.cookies.get({
-          url: targetUrl,
-          name: cookieName,
-        });
-        const cookie = cookies[0] ?? null;
-        writeBrowserServiceJson(response, 200, {
-          ok: true,
-          cookie: cookie
-            ? {
-                name: cookie.name || "",
-                value: cookie.value || "",
-                domain: cookie.domain || "",
-                path: cookie.path || "/",
-                secure: Boolean(cookie.secure),
-                httpOnly: Boolean(cookie.httpOnly),
-                session: Boolean(cookie.session),
-                sameSite: cookie.sameSite || "unspecified",
-                expirationDate:
-                  typeof cookie.expirationDate === "number" &&
-                  Number.isFinite(cookie.expirationDate)
-                    ? cookie.expirationDate
-                    : null,
-              }
-            : null,
-        });
-      } catch (error) {
-        writeBrowserServiceJson(response, 400, {
-          error:
-            error instanceof Error
-              ? error.message
-              : "Failed to set browser cookie.",
-        });
-      }
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/tabs") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const targetUrl =
-        typeof payload.url === "string" && payload.url.trim()
-          ? payload.url.trim()
-          : HOME_URL;
-      const background = payload.background === true;
-      const workspace = await ensureTargetBrowserSpace("tabs:create");
-      const tabSpace = browserTabSpaceState(workspace, targetSpace, ensuredSessionId, {
-        createIfMissing: targetSpace === "agent" && Boolean(requestedSessionId),
-        useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId,
-      });
-      if (!workspace) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser workspace is available.",
-        });
-        return;
-      }
-
-      const nextTabId = createBrowserTab(targetWorkspaceId, {
-        url: targetUrl,
-        browserSpace: targetSpace,
-        sessionId: ensuredSessionId,
-      });
-      if (!nextTabId) {
-        writeBrowserServiceJson(response, 500, {
-          error: "Failed to create browser tab.",
-        });
-        return;
-      }
-
-      if (!background && tabSpace) {
-        tabSpace.activeTabId = nextTabId;
-        if (targetWorkspaceId === activeBrowserWorkspaceId) {
-          updateAttachedBrowserView();
-        }
-      }
-
-      emitBrowserState(targetWorkspaceId, targetSpace);
-      await persistBrowserWorkspace(targetWorkspaceId);
-      writeBrowserServiceJson(
-        response,
-        200,
-        browserWorkspaceSnapshot(targetWorkspaceId, targetSpace, ensuredSessionId, {
-          useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId,
-        }),
-      );
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/evaluate") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const expression =
-        typeof payload.expression === "string" ? payload.expression.trim() : "";
-      if (!expression) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Field 'expression' is required.",
-        });
-        return;
-      }
-
-      const workspace = await ensureTargetBrowserSpace("evaluate");
-      if (!workspace) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-
-      const result =
-        await activeTab.view.webContents.executeJavaScript(expression);
-      writeBrowserServiceJson(response, 200, {
-        tabId: activeTab.state.id,
-        result: serializeBrowserEvalResult(result),
-      });
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/context-click") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const x =
-        typeof payload.x === "number" && Number.isFinite(payload.x)
-          ? Math.round(payload.x)
-          : NaN;
-      const y =
-        typeof payload.y === "number" && Number.isFinite(payload.y)
-          ? Math.round(payload.y)
-          : NaN;
-      if (!Number.isFinite(x) || x < 0 || !Number.isFinite(y) || y < 0) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Fields 'x' and 'y' must be non-negative numbers.",
-        });
-        return;
-      }
-
-      const workspace = await ensureTargetBrowserSpace("context-click");
-      if (!workspace) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-
-      if (mainWindow && !mainWindow.isDestroyed() && !mainWindow.isFocused()) {
-        mainWindow.focus();
-      }
-      await withProgrammaticBrowserInput(activeTab.view.webContents, async () => {
-        activeTab.view.webContents.focus();
-        await activeTab.view.webContents.sendInputEvent({
-          type: "mouseMove",
-          x,
-          y,
-        });
-        await activeTab.view.webContents.sendInputEvent({
-          type: "mouseDown",
-          x,
-          y,
-          button: "right",
-          clickCount: 1,
-        });
-        await activeTab.view.webContents.sendInputEvent({
-          type: "mouseUp",
-          x,
-          y,
-          button: "right",
-          clickCount: 1,
-        });
-      });
-
-      writeBrowserServiceJson(response, 200, {
-        ok: true,
-        tabId: activeTab.state.id,
-        x,
-        y,
-      });
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/mouse") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const x =
-        typeof payload.x === "number" && Number.isFinite(payload.x)
-          ? Math.round(payload.x)
-          : NaN;
-      const y =
-        typeof payload.y === "number" && Number.isFinite(payload.y)
-          ? Math.round(payload.y)
-          : NaN;
-      const action =
-        payload.action === "double_click" || payload.action === "hover"
-          ? payload.action
-          : "click";
-      if (!Number.isFinite(x) || x < 0 || !Number.isFinite(y) || y < 0) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Fields 'x' and 'y' must be non-negative numbers.",
-        });
-        return;
-      }
-
-      const workspace = await ensureTargetBrowserSpace("mouse");
-      if (!workspace) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-
-      if (mainWindow && !mainWindow.isDestroyed() && !mainWindow.isFocused()) {
-        mainWindow.focus();
-      }
-      await withProgrammaticBrowserInput(activeTab.view.webContents, async () => {
-        activeTab.view.webContents.focus();
-        await activeTab.view.webContents.sendInputEvent({
-          type: "mouseMove",
-          x,
-          y,
-        });
-
-        if (action === "click" || action === "double_click") {
-          await activeTab.view.webContents.sendInputEvent({
-            type: "mouseDown",
-            x,
-            y,
-            button: "left",
-            clickCount: 1,
-          });
-          await activeTab.view.webContents.sendInputEvent({
-            type: "mouseUp",
-            x,
-            y,
-            button: "left",
-            clickCount: 1,
-          });
-        }
-        if (action === "double_click") {
-          await activeTab.view.webContents.sendInputEvent({
-            type: "mouseDown",
-            x,
-            y,
-            button: "left",
-            clickCount: 2,
-          });
-          await activeTab.view.webContents.sendInputEvent({
-            type: "mouseUp",
-            x,
-            y,
-            button: "left",
-            clickCount: 2,
-          });
-        }
-      });
-
-      writeBrowserServiceJson(response, 200, {
-        ok: true,
-        tabId: activeTab.state.id,
-        action,
-        x,
-        y,
-      });
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/keyboard") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const action = payload.action === "press" ? "press" : "insert_text";
-      const text = typeof payload.text === "string" ? payload.text : "";
-      const key =
-        typeof payload.key === "string" && payload.key.trim()
-          ? payload.key.trim()
-          : "";
-      const clear = payload.clear === true;
-      const submit = payload.submit === true;
-      if (action === "press" && !key) {
-        writeBrowserServiceJson(response, 400, {
-          error: "Field 'key' is required for keyboard press actions.",
-        });
-        return;
-      }
-
-      const workspace = await ensureTargetBrowserSpace("keyboard");
-      if (!workspace) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-
-      if (mainWindow && !mainWindow.isDestroyed() && !mainWindow.isFocused()) {
-        mainWindow.focus();
-      }
-      await withProgrammaticBrowserInput(activeTab.view.webContents, async () => {
-        activeTab.view.webContents.focus();
-        if (action === "press") {
-          await sendBrowserKeyPress(activeTab.view.webContents, key);
-          return;
-        }
-        if (clear) {
-          await clearFocusedBrowserTextInput(activeTab.view.webContents);
-        }
-        if (text) {
-          await activeTab.view.webContents.insertText(text);
-        }
-        if (submit) {
-          await sendBrowserKeyPress(activeTab.view.webContents, "Enter");
-        }
-      });
-
-      writeBrowserServiceJson(response, 200, {
-        ok: true,
-        tabId: activeTab.state.id,
-        action,
-        text_length: action === "insert_text" ? text.length : 0,
-        key: action === "press" ? key : "",
-        clear,
-        submit,
-      });
-      return;
-    }
-
-    if (method === "POST" && pathname === "/api/v1/browser/screenshot") {
-      const payload = await readBrowserServiceJsonBody(request);
-      const workspace = await ensureTargetBrowserSpace("screenshot");
-      if (!workspace) {
-        return;
-      }
-      const activeTab = getActiveBrowserTab(
-        targetWorkspaceId,
-        targetSpace,
-        ensuredSessionId,
-        { useVisibleAgentSession: targetSpace === "agent" && !requestedSessionId },
-      );
-      if (!activeTab) {
-        writeBrowserServiceJson(response, 409, {
-          error: "No active browser tab is available.",
-        });
-        return;
-      }
-
-      const format = payload.format === "jpeg" ? "jpeg" : "png";
-      const qualityRaw =
-        typeof payload.quality === "number" ? payload.quality : 90;
-      const quality = Math.max(0, Math.min(100, Math.round(qualityRaw)));
-      const image = await activeTab.view.webContents.capturePage();
-      const buffer = format === "jpeg" ? image.toJPEG(quality) : image.toPNG();
-      const size = image.getSize();
-
-      writeBrowserServiceJson(response, 200, {
-        tabId: activeTab.state.id,
-        mimeType: format === "jpeg" ? "image/jpeg" : "image/png",
-        width: size.width,
-        height: size.height,
-        base64: buffer.toString("base64"),
-      });
-      return;
-    }
-
-    writeBrowserServiceJson(response, 404, { error: "Not found." });
-  } catch (error) {
-    writeBrowserServiceJson(response, 500, {
-      error:
-        error instanceof Error
-          ? error.message
-          : "Browser service request failed.",
-    });
-  }
-}
+// handleDesktopBrowserServiceRequest moved to browser-pane/tab-state.ts.
 
 async function startDesktopBrowserService(): Promise<void> {
   if (desktopBrowserServiceServer) {
@@ -6434,7 +5460,7 @@ async function startDesktopBrowserService(): Promise<void> {
 
   const authToken = randomUUID();
   const server = createServer((request, response) => {
-    void handleDesktopBrowserServiceRequest(request, response);
+    void browserHttpService.handleRequest(request, response);
   });
 
   await new Promise<void>((resolve, reject) => {
@@ -17393,6 +16419,67 @@ async function closeBrowserTab(
   await ensureBrowserWorkspace(options.workspaceId, browserSpace, normalizedSessionId);
   return closeBrowserTabInner(tabId, options);
 }
+
+// Renderer-→runtime browser HTTP service. The route handler lives in
+// browser-pane/http-service.ts; here we just wire deps. Server lifecycle
+// (start/stop, auth-token rotation, capability config sync) stays in
+// main.ts.
+const browserHttpService: BrowserHttpService = createBrowserHttpService({
+  getMainWindow: () => mainWindow,
+  getActiveWorkspaceId: () => activeBrowserWorkspaceId,
+  getAuthToken: () => desktopBrowserServiceAuthToken,
+  homeUrl: HOME_URL,
+  browserSpaceId: (value, fallback) => browserSpaceId(value, fallback),
+  ensureBrowserWorkspace: (workspaceId, space, sessionId) =>
+    ensureBrowserWorkspace(workspaceId, space, sessionId) as unknown as ReturnType<
+      Parameters<typeof createBrowserHttpService>[0]["ensureBrowserWorkspace"]
+    >,
+  ensureUserBrowserLock: (workspaceId, sessionId, reason) => {
+    const result = ensureUserBrowserLock(workspaceId, sessionId, reason);
+    return result.ok
+      ? { ok: true }
+      : { ok: false, lockHolderSessionId: result.lockHolderSessionId };
+  },
+  touchAgentSessionBrowserSpace: (workspaceId, sessionId) =>
+    touchAgentSessionBrowserSpace(workspaceId, sessionId),
+  browserWorkspaceSnapshot: (workspaceId, space, sessionId, options) =>
+    browserWorkspaceSnapshot(workspaceId, space, sessionId, options),
+  browserTabSpaceState: (workspace, space, sessionId, options) =>
+    browserTabSpaceState(
+      workspace as unknown as BrowserWorkspaceState | null | undefined,
+      space,
+      sessionId,
+      options,
+    ) as unknown as ReturnType<
+      Parameters<typeof createBrowserHttpService>[0]["browserTabSpaceState"]
+    >,
+  getActiveBrowserTab: (workspaceId, space, sessionId, options) =>
+    getActiveBrowserTab(workspaceId, space, sessionId, options) as unknown as ReturnType<
+      Parameters<typeof createBrowserHttpService>[0]["getActiveBrowserTab"]
+    >,
+  syncBrowserState: (workspaceId, tabId, space, sessionId) =>
+    syncBrowserState(workspaceId, tabId, space, sessionId),
+  navigateActiveBrowserTab: (workspaceId, targetUrl, space, sessionId) =>
+    navigateActiveBrowserTab(workspaceId, targetUrl, space, sessionId),
+  createBrowserTab: (workspaceId, options) =>
+    createBrowserTab(workspaceId, options),
+  setActiveBrowserTab: (tabId, options) => setActiveBrowserTab(tabId, options),
+  closeBrowserTab: (tabId, options) => closeBrowserTab(tabId, options),
+  updateAttachedBrowserView: () => updateAttachedBrowserView(),
+  emitBrowserState: (workspaceId, space) => emitBrowserState(workspaceId, space),
+  persistWorkspace: (workspaceId) => persistBrowserWorkspace(workspaceId),
+  emitWorkbenchOpenBrowser: (payload) => emitWorkbenchOpenBrowser(payload),
+  operatorSurfaceContextPayload: (workspaceId) =>
+    operatorSurfaceContextPayload(workspaceId),
+  browserPagePayload: (tab) =>
+    browserPagePayload(tab as unknown as BrowserTabRecord),
+  withProgrammaticBrowserInput: (webContents, callback) =>
+    withProgrammaticBrowserInput(webContents, callback),
+  sendBrowserKeyPress: (webContents, keyCode, modifiers) =>
+    sendBrowserKeyPress(webContents, keyCode, modifiers),
+  clearFocusedBrowserTextInput: (webContents) =>
+    clearFocusedBrowserTextInput(webContents),
+});
 
 const TEXT_FILE_EXTENSIONS = new Set([
   ".txt",
